requirements.txt
clarifai-grpc
clarifai
python-datamuse
nltk
numpy